version 2.1

+++ Version2_1/grind/models.py
@@ -12,9 +12,10 @@ class CacheFile(models.Model):
     pc = models.FileField(upload_to='pc_cache/', default='DEFAULT VALUE', blank=True, null=True)
     
 class Point_Cloud(models.Model):
-    x = models.IntegerField(default=0)
-    y = models.IntegerField(default=0)
-    z = models.IntegerField(default=0)
+    x = models.FloatField(default=0)
+    y = models.FloatField(default=0)
+    z = models.FloatField(default=0)
     imp = models.FloatField(default=0)
 #    name = models.CharField(max_length=200, null=True)
     xypoint = django.contrib.gis.db.models.GeometryField(srid=28992, null=True)
+#    xyzpoint = django.contrib.gis.db.models.GeometryField(srid=28992, null=True)
